requirements_gpu.txt
numpy
tensorflow_addons>=0.9.1
tensorflow-gpu>=2.0
